complete refactor of results.py for better modular organization
